Set elements to be hidden while signed in/out, also set timeouts for success and error messages
Arranged Sign In and Out headers to be inside their respective forms, so
that they can be hidden if the user is signed in or out. Set the sign
out button to be hidden while the user is not signed in. Set the sign
in and up forms to be hidden once the user signs in and the sign out
button to show as well. Created timeout functions for all the success
and error messages.